refactor of admin/user apis
