+++ src/api/index.php
@@ -72,7 +72,10 @@ require __DIR__ . '/routes/geocoder.php';
 require __DIR__ . '/routes/kiosks.php';
 require __DIR__ . '/routes/email/mailchimp.php';
 require __DIR__ . '/routes/search.php';
-require __DIR__ . '/routes/users/users-admin.php';
+
+//user
+require __DIR__ . '/routes/users/user.php';
+require __DIR__ . '/routes/users/user-admin.php';
 require __DIR__ . '/routes/users/user-current.php';
 require __DIR__ . '/routes/users/user-settings.php';
